tests/kms_modes_available_crc: Fix NV12 failure
The size of the UV plane is not calculated correctly - height is not tile
aligned. Make use of the stride and offset values intitialized in the
previous patch to calculate plane size. The next step would be to rewrite
the test to make use of library functions, but for now this should fix
NV12.

Cc: Juha-Pekka Heikkila <juhapekka.heikkila@gmail.com>
Signed-off-by: Dhinakaran Pandiyan <dhinakaran.pandiyan@intel.com>
Reviewed-by: José Roberto de Souza <jose.souza@intel.com>