tests/testdisplay: Speed up testdisplay for ci
No point in sleeping 5 seconds between each modeset when running in ci.
So let's just make the default sleep interval 0 seconds and leave it up
to the user to bump it via -s when running interactively.

On a KBL + LSPCON + a 4k HDMI display with 36 modes listed:
time ./tests/testdisplay
- real	3m12,026s
- user	0m1,920s
- sys	0m1,241s
+ real	0m14,681s
+ user	0m3,135s
+ sys	0m1,340s

Cc: Martin Peres <martin.peres@linux.intel.com>
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>
Reviewed-by: Chris Wilson <chris@chris-wilson.co.uk>