lib/igt_kms: Don't reset VRR_ENABLED on every commit
If the VRR tests failed then there was a chance that VRR could be
left enabled when exiting the test so the VRR_ENABLED property was
reset to 0 whenever the pipe was reset.

However, in doing so the pipe's state was considered changed even
if VRR_ENABLED was already 0. This causes the pipe to be added to
commits where it previously wasn't.

Other tests with properties that can persist for failures (like color
management) don't bother resetting the properties - so this patch
changes VRR_ENABLED to work the same.

Cc: Harry Wentland <harry.wentland@amd.com>
Cc: Petri Latvala <petri.latvala@intel.com>
Cc: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=109490
Signed-off-by: Nicholas Kazlauskas <nicholas.kazlauskas@amd.com>
Reviewed-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Reviewed-by: Harry Wentland <harry.wentland@amd.com>