lib/igt_fb: Fix merge of preliminary patches for clearing YUV
Preliminary patches for the YUV testing series were broken without
subsequent patches that fix them when moving things around.

The argument provided to clear_yuv_buffer should be the framebuffer,
not the DRM file descriptor (fb vs fd). Also, an assert was added to
clear_yuv_buffer to ensure that it's only called for YUV formats,
which has to be reflected in the calling function.

Signed-off-by: Paul Kocialkowski <paul.kocialkowski@bootlin.com>
Reviewed-by: Maxime Ripard <maxime.ripard@bootlin.com>