README: Add pointer to Dockerfiles for easy lookup of dependencies
I spent quite some time trying to install all the dependencies to build
igt in for my distribution. After installing the packages, I noticed
that the Dockerfiles had the exact packages required for each
distribution. Add a pointer to the Dockerfiles in the README to save
folks time on setting up the environment required to build igt.

CC: Rodrigo Vivi <rodrigo.vivi@intel.com>
Signed-off-by: Easwar Hariharan <easwar.hariharan@intel.com>
Reviewed-by: Rodrigo Vivi <rodrigo.vivi@intel.com>