intel-ci: Disable i915_suspend@shrink
This test produces an awful, awful lot of redundant output as it tries
to find just the right amount of memory pressure to cause an
out-of-memory event in the middle of suspend. That is always quite a
slow process, taking 90s on a normal machine and 500+s on skl-y.
Furthermore, even when we do achieve the perfect setup, the test
frequently locks up and fails to resume with no indication that it is a
bug in the driver. The shrinker and oomkiller (plus i915) do not make for
a pleasant time!

Enough of Martin's whinging, I see no way of easily making this test
quieter, quicker and more efficacious, relegate it to the masochist only
stable.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Martin Peres <martin.peres@free.fr>
Cc: Petri Latvala <petri.latvala@intel.com>
Reviewed-by: Martin Peres <martin.peres@free.fr>