autotools: Always build VC4 helpers with the library
With recent changes to the igt_fb code, VC4 support is now integrated
in common helpers. As a result, VC4 helpers have to be built as part
of the library.

This is already the case with meson, but VC4 support was removed from
autotools. Bring it back unconditionally to match meson behaviour.

This fixes a build error with autotools.

Signed-off-by: Paul Kocialkowski <paul.kocialkowski@bootlin.com>