tests/kms_plane: Keep testing pixel formats even if one fails
In the pixel-format-pipe-*-planes* family of subtests currently we exit
early if there is any CRC mismatch, leaving all the remaining formats
untested.

Let's fix that by moving the assert till after all the iterations.

Also log each mismatch with its context in a single line so that it's
easy to look for.

Cc: Daniel Vetter <daniel@ffwll.ch>
Reviewed-by: Martin Peres <martin.peres@linux.intel.com>
Signed-off-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>