tests/kms_chamelium: fix typo in check_analog_bridge
This function checks whether a port is an analog bridge. Analog
bridges provide their own EDID thus the vendor should be different
from "IGT".

There was a typo that made the function always return true for VGA
ports. The condition was essentially:

    thing != 'I' || thing != 'T'

which is always true.

Apparently this didn't cause any harm, but it's probably better to fix
it anyway in case some tests were skipped and shouldn't.

Signed-off-by: Simon Ser <simon.ser@intel.com>
Acked-by: Antonio Argenziano <antonio.argenziano@intel.com>