runner: Reinitialize compiled dmesg regexp each parsing session
Which regexp gets compiled is settings specific, depending whether we
run piglit-style or not.

If it's optimized to be initialized only once and it is a global
variable, it will be "stuck" in the mode we have selected with the first
run, which may break tests.

Let's remove this optimization and initialize it each time, as it takes
less 0.002s on my hardware.

Signed-off-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>
Reviewed-by: Simon Ser <simon.ser@intel.com>