tests/kms_plane_scaling: Check supported format for rotation
Let's check if 90/270 rotation is supported for the format that we are
trying to test.

Currently, if we try to test unsupported format the kernel complains with
the following message.

"[drm:skl_plane_check [i915]] Unsupported pixel format Y210
little-endian (0x30313259) for 90/270!"

v2: Use igt_plane_has_format_mod() function to select formats that are
    capable of 90/270 rotation (Ville)

Suggested-by: Ville Syrjälä <ville.syrjala@linux.intel.com>
Signed-off-by: Mika Kahola <mika.kahola@intel.com>
Reviewed-by: Ville Syrjälä <ville.syrjala@linux.intel.com>