intel-ci: Sanitycheck CRC readback on the first pipe only
For fast-feedback, we need to sanitycheck that CRC reads work, otherwise
the display validation tests are flawed. However, we don't need to
extensively test each pipe as the HW is fundamentally the same. Instead,
do the full suite of tests on one pipe, and just verify that the interface
works for the other pipes. This should speed BAT up by reducing
NUM_PIPES suspend and resume cycles to just 1.

The shards should cover the extended test suite on all pipes for
complete validation.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Petri Latvala <petri.latvala@intel.com>
Cc: Martin Peres <martin.peres@free.fr>
Cc: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Acked-by: Martin Peres <martin.peres@free.fr>
Reviewed-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>